# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DisClone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── SERVERS ─────────────────────────────────────────────────────────────────
  /servers:
    get:
      operationId: listMyServers
      tags: [servers]
      summary: List servers the current user belongs to
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Create a new server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerInput"
      responses:
        "201":
          description: Created server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /servers/discover:
    get:
      operationId: discoverServers
      tags: [servers]
      summary: Discover public servers
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Public servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
  /servers/{serverId}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get server details
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
    patch:
      operationId: updateServer
      tags: [servers]
      summary: Update server settings
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpdate"
      responses:
        "200":
          description: Updated server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /servers/{serverId}/leave:
    post:
      operationId: leaveServer
      tags: [servers]
      summary: Leave a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /servers/{serverId}/summary:
    get:
      operationId: getServerSummary
      tags: [servers]
      summary: Get server summary stats (members online, total messages, boost level)
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerSummary"

  # ─── CHANNELS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, username, discriminator, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        discriminator:
          type: string
        displayName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, idle, dnd, invisible, offline]
        customStatus:
          type: ["string", "null"]
        customStatusEmoji:
          type: ["string", "null"]
        isNitro:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── SERVER ───────────────────────────────────────────────────────────────
    Server:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        ownerId:
          type: integer
        isPublic:
          type: boolean
        memberCount:
          type: integer
        boostLevel:
          type: integer
        boostCount:
          type: integer
        vanityUrl:
          type: ["string", "null"]
        verificationLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
    ServerDetail:
      type: object
      required: [id, name, ownerId, createdAt, channels, roles]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        ownerId:
          type: integer
        isPublic:
          type: boolean
        memberCount:
          type: integer
        boostLevel:
          type: integer
        boostCount:
          type: integer
        vanityUrl:
          type: ["string", "null"]
        verificationLevel:
          type: integer
        channels:
          type: array
          items:
            $ref: "#/components/schemas/Channel"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
    ServerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
        iconUrl:
          type: string
        isPublic:
          type: boolean
    ServerUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        isPublic:
          type: boolean
        vanityUrl:
          type: ["string", "null"]
        verificationLevel:
          type: integer
    ServerSummary:
      type: object
      required: [serverId, onlineCount, totalMembers, totalMessages, boostLevel, boostCount]
      properties:
        serverId:
          type: integer
        onlineCount:
          type: integer
        totalMembers:
          type: integer
        totalMessages:
          type: integer
        boostLevel:
          type: integer
        boostCount:
          type: integer

    # ─── CHANNEL ──────────────────────────────────────────────────────────────
    Channel:
      type: object
      required: [id, serverId, name, type, position, createdAt]
      properties:
        id:
          type: integer
        serverId:
          type: integer
        parentId:
          type: ["integer", "null"]
        name:
          type: string
        topic:
          type: ["string", "null"]
        type:
          type: string
          enum: [text, voice, stage, forum, announcement, category]
        position:
          type: integer
        slowmode:
          type: integer
        nsfw:
          type: boolean
        userLimit:
          type: ["integer", "null"]
        bitrate:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    Role:
      type: object
      required: [id, serverId, name, color, position, createdAt]
      properties:
        id:
          type: integer
        serverId:
          type: integer
        name:
          type: string
        color:
          type: string
        position:
          type: integer
        hoist:
          type: boolean
        mentionable:
          type: boolean
        permissions:
          type: integer
        createdAt:
          type: string
          format: date-time
...