# lib/api-zod/src/generated/types/serverDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisClone API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Channel } from './channel';
import type { Role } from './role';

export interface ServerDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  ownerId: number;
  isPublic?: boolean;
  memberCount?: number;
  boostLevel?: number;
  boostCount?: number;
  /** @nullable */
  vanityUrl?: string | null;
  verificationLevel?: number;
  channels: Channel[];
  roles: Role[];
  createdAt: Date;
}