# lib/api-zod/src/generated/types/serverInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisClone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServerInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  description?: string;
  iconUrl?: string;
  isPublic?: boolean;
}