# lib/api-zod/src/generated/types/serverSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisClone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServerSummary {
  serverId: number;
  onlineCount: number;
  totalMembers: number;
  totalMessages: number;
  boostLevel: number;
  boostCount: number;
}