# lib/api-zod/src/generated/types/serverUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisClone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServerUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  isPublic?: boolean;
  /** @nullable */
  vanityUrl?: string | null;
  verificationLevel?: number;
}